Extracted DopeDatabase from WrappedApolloProvider
